add chatbot

+++ WebApiShope/Repositories/StatusesReposetory.cs
@@ -17,7 +17,7 @@ namespace Repositories
             this._DBContext = DBContext;
         }
 
-        async public Task<Status?> GetStatusByID(int statusID)
+        async public Task<Status?> GetStatusByID(long statusID)
         {
             return await _DBContext.Statuses.FirstOrDefaultAsync(x => x.StatusId == statusID);
         }
